# PERF: restore DatetimeIndex.__iter__ performance by using non-EA implementation