@jbrockmendel 2019-06-07
LGTM.  Only question is if in principle we should be wrapping `self._data.__iter__`, but I don't think its worth the hassle.