@TomAugspurger 2019-06-07
Thanks. In theory, a similar chunking trick could be done for PeriodArray & TimedeltaArray, right?  The only thing to change would be the function doing the conversion in

```python
            converted = tslib.ints_to_pydatetime(data[start_i:end_i],
                                                 tz=self.tz, freq=self.freq,
                                                 box="timestamp")
 ```

Happy to leave that as a followup.