@TomAugspurger 2019-06-07
No worries, I opened https://github.com/pandas-dev/pandas/issues/26713 to
track this.

On Fri, Jun 7, 2019 at 2:21 PM Christopher Whelan <notifications@github.com>
wrote:

> @jreback <https://github.com/jreback> Thanks for merging!
>
> @TomAugspurger <https://github.com/TomAugspurger> Yes, definitely a
> possibility. There's probably mid-single-digit factor speedups available
> there. Unfortunately, I'm headed to Europe for vacation next week and
> starting a new job shortly thereafter so can't promise to make any headway
> on those.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/pandas-dev/pandas/pull/26702?email_source=notifications&email_token=AAKAOITHGCNLCEIOHQRJW5TPZKYMNA5CNFSM4HVT2QQ2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGODXGYLQI#issuecomment-500008385>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAKAOIRY3W43FBBZU7WWP3LPZKYMNANCNFSM4HVT2QQQ>
> .
>